fix(site): fix bad render/crawl of site
Was gettings infernal error for hours:
[0] Error: EISDIR: illegal operation on a directory, open '/Users/jaketrent/dev/roboto2/docs'
[0] webpack --watch --config ./build/webpack.config.babel.js exited with code 1

Finally git-bisected to glory and found the bad commit:
01d619f6aeea14b9ac7566de3b38252281e2d341

Readded every file and npm started between.  Because of the bad error,
don't have a particular line of code, but the smoking gun is in the
hands of:
https://github.com/markdalgleish/static-site-generator-webpack-plugin

My guess is "?broken" was trying to create a "docs/?broken" directory
when crawling the site, which is apparently problematic.