fix(card): make icon correct new size
Embracing the icon dependency.  For the card, this seems to make sense.
For others, like button, I'm less sure.  If you have a card, you
probably have actions.  If you have a button, you might not use the
icons and still have a wide range of use cases possible to you.  Might
not be worth adding the Icon dependency weight in that case.  If we had
these sizes importable differently, we could just import the sizes
object and not the constructor function, which has the static member of
sizes today.  We did the ctor static attachment for ease of use for the
developer.  Perhaps we could do both?  Multiple export modes would be cheap.