fix(avatar): hide img or initials based on img load state
Help remove color halo that bled out from behind loaded images.  Had to
change the gravatar fallback strategy to support this.  Otherwise, the
rule was that we always show initials in the case of gravatar images.
But we don't want this, because depending on the color combo, the halo
is visible.  Instead, now we fail gravatar images like any other.

Closes #191