fix(site): create new stacking context for text input example
The pseudo elements for the text input come behind the input field only if a new
stacking context is made for them (position absolute + z-index non-auto).  But
because of the negative z-index, they are showing up apparently behind the
parent stacking context (white background of html element???).  To solve this,
I'm creating a new stacking context for the react example output.  I'm not fully
groking why this is working.  But it seems that the closer/intermediate stacking
context is required.  This will need to be of note to developers who use text
input.  It is probably something we'll want to build into our form layout
component as well.