fix(prop-types): workaround circular json error on array validations
When the validator receives a prop that happens to be a circular
json, JSON.stringify will throw an error, and as stated on
prop-types docs you can't throw inside a PropTypes.oneOf validator.

It happened to me while using `ActionMenu` with multiple
`ActionMenu.Item`, the children was an array of elements and failed
on stringify due to a circular json.
In my specific case, the `PropTypes.oneOf` will validate as ok if
the stringify don't throw.

Probably the message `with value ${JSON.stringify(...)}` is not that
pretty for this circular json situation (you can test it by passing
multiple ActionMenu.Item and a non ActionMenu.Item to an ActionMenu).
But this case (no match on the `PropTypes.oneOf` is not happening to
me, so I'm not planning to address it right now.