fix(emptystate): set should not update after unmounting component
There was an issue where state was being updated if the component was quickly mounted and unmounted.
This change will not update state if the component has been unmounted. also I converted to react
hooks and used an effect closure to make it easier to reason about

BREAKING CHANGE: react version upgrade