fix(tab): properly calculate nowrap slider length
Unsolved mystery: setting flex-shrink: 0 on the list-item was not working
well with the getBoundingClientRect() width measurement.  The width
remained the same as when flex-shrink was 1.  Odd.

Also had to add another text inner div to allow super long text to not
push the text outside the 64px height.  Some combo in the adjusted css
didn't jive well with the change.