fix(collapsible): animation & height issues
- Fixes close animations not working. The bug was caused by the height
not being set to the calculated auto height before being set to zero. Fixed
by not setting height to 0 before calculating the height.
- Fixes the container not resizing with its content when it starts open. The
bug was caused by the height not being set to "auto" when it starts opened.
Fixed by resolving waitForHeightTransitionToEnd() if container is already at
target height.
- Fixes the drawer breaking if user opens while it's closing or vice versa.
The bug was caused by not checking if the collapse was in the expected
state after the animation finishes. Fixed by checking the measured height
of the element after the animation finished and only doing close/open cleanup
code if the container is closed/opened.