fix(navuser): avoid swallowing click
The click being swallowed was in a very specific instance, as seen by
the added story. Still not sure why the onClick was not being triggered.
Before this commit, would take an extra click (click navuser to click
off and closer actionmenu due to document close listeners, then another
click to open the navuser). The change makes all that happen in one
click.

Resolves: #1656