refactor(theme): import react as cjs
Named imports can not be used when importing CJS from inside ESM,
which our components now are.

These are not allowed:

- import { createContext } from 'react'
- import * as React from 'react'

Only the default import is available:

- import React from 'react'

Some tooling is more ESM friendly than others and seems to overlook
this. Typescript/tsc itself, for instance.  Jest, in test, still
seems to not be super flexible with ESM. I've seen the same in
Storybook.

The symptom is that in other packages (eg, badge), when storyshots are
run, there are no stories found because the (eg) theme dependency
evaluates some React-related code and fails.  Eg,

import React, { createContext } from  'react'

// ...

createContext()

Reports that createContext is not a function, fails and thus storyshots
doesn't find a working story.

There are probably other tooling-related solutions. But, looking at
React, it definitely only has CJS and UMD builds, even at version 17.

So for now, until tooling catches up, I recommend we use the default
export when importing react and continue to treat it like the CJS module
it is. This will be a little painful because of the code changes
required.

Some clear rules about the CJS/ESM interop:
https://redfin.engineering/node-modules-at-war-why-commonjs-and-es-modules-cant-get-along-9617135eeca1