refactor(examples): kitchen-sink craco to look for js
The rules totally overlap, but I think this is a more accurate naming
conceptually.

It is the modules being IMPORTED that need to have the module field
looked at instead of the main. This solves the problem of webpack not
seeing libs like dayzed and date-fns as dual-published esm/cjs packages.