refactor(core): don't treat css as modules
The CSS is mostly vars anyway. And they're mostly put in :root anyway.
This takes away a false sense of modularity.

The exception is accessibility.

Considering the components don't use modules and instead just use BEM, I
thought that it might be more consistent in dev and dev tooling to opt
out of modules here too.

BREAKING CHANGE: css is no longer in modules. Was misleading.