fix(normalize): revert core back to camelCase variables
Issue is that this normalize package didn't receive a breaking change,
but had it starting using core@9, which changed to dashified css vars.
These vars are in the global :root and used in this package. But we
don't want to ship that var change to people until they're ready. This
has probably been breaking pages (hopefully subtly?!) that have looser
version dependency on the normalize package.

After this, we want to go back to dashified, use core@9, do a major
version bump and then patch fix all latest components to peerDepend on
the latest normalize.

Related to #1947