fix(drawer): make keyboard accessible by tabbing
Followed this ARIA according example:
https://www.w3.org/TR/wai-aria-practices-1.1/examples/accordion/accordion.html

Didn't do arrowing. Drawer implementation today doesn't have a
collection or a stack concept. The treatment of those as a group
wouldn't be super nice given that present fact. I'll count this a nice
improvement over the lack of even focus ability to this point.

Resolves: #1836